+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 v0.3.0 (in development)
 -----------------------
 - Support Python 3.9
+- Gave `attrrollback()` and `itemrollback()` new arguments `copy` and
+  `deepcopy` for storing copies of the specified attribute/item
 
 v0.2.0 (2020-10-19)
 -------------------